Move all internal types from snaps-types to snaps-utils
