+++ packages/snaps-execution-environments/src/common/validation.ts
@@ -1,5 +1,4 @@
-import { SnapKeyring } from '@metamask/snaps-types';
-import { ChainIdStruct, HandlerType } from '@metamask/snaps-utils';
+import { SnapKeyring, ChainIdStruct, HandlerType } from '@metamask/snaps-utils';
 import {
   assertStruct,
   Json,
