+++ packages/snaps-utils/package.json
@@ -52,6 +52,7 @@
   "dependencies": {
     "@babel/core": "^7.18.6",
     "@babel/types": "^7.18.7",
+    "@metamask/providers": "^10.2.1",
     "@metamask/snaps-types": "^0.26.2",
     "@metamask/utils": "^3.3.1",
     "@noble/hashes": "^1.1.3",
